Remove New suffix
